+++ pydantic/class_validators.py
@@ -3,10 +3,10 @@ from dataclasses import dataclass
 from enum import IntEnum
 from itertools import chain
 from types import FunctionType
-from typing import Callable, Dict
+from typing import Any, Callable, Dict, List, Optional, Set
 
 from .errors import ConfigError
-from .utils import in_ipython
+from .utils import AnyCallable, in_ipython
 
 
 class ValidatorSignature(IntEnum):
@@ -18,17 +18,19 @@ class ValidatorSignature(IntEnum):
 
 @dataclass
 class Validator:
-    func: Callable
+    func: AnyCallable
     pre: bool
     whole: bool
     always: bool
     check_fields: bool
 
 
-_FUNCS = set()
+_FUNCS: Set[str] = set()
 
 
-def validator(*fields, pre: bool = False, whole: bool = False, always: bool = False, check_fields: bool = True):
+def validator(
+    *fields: str, pre: bool = False, whole: bool = False, always: bool = False, check_fields: bool = True
+) -> Callable[[AnyCallable], classmethod]:
     """
     Decorate methods on the class indicating that they should be used to validate fields
     :param fields: which field(s) the method should be called on
@@ -45,7 +47,7 @@ def validator(*fields, pre: bool = False, whole: bool = False, always: bool = Fa
             "E.g. usage should be `@validator('<field_name>', ...)`"
         )
 
-    def dec(f):
+    def dec(f: AnyCallable) -> classmethod:
         # avoid validators with duplicated names since without this validators can be overwritten silently
         # which generally isn't the intended behaviour, don't run in ipython - see #312
         if not in_ipython():  # pragma: no branch
@@ -54,26 +56,30 @@ def validator(*fields, pre: bool = False, whole: bool = False, always: bool = Fa
                 raise ConfigError(f'duplicate validator function "{ref}"')
             _FUNCS.add(ref)
         f_cls = classmethod(f)
-        f_cls.__validator_config = fields, Validator(f, pre, whole, always, check_fields)
+        f_cls.__validator_config = fields, Validator(f, pre, whole, always, check_fields)  # type: ignore
         return f_cls
 
     return dec
 
 
+ValidatorListDict = Dict[str, List[Validator]]
+
+
 class ValidatorGroup:
-    def __init__(self, validators):
-        self.validators: Dict[str, Validator] = validators
+    def __init__(self, validators: ValidatorListDict) -> None:
+        self.validators = validators
         self.used_validators = {'*'}
 
-    def get_validators(self, name):
+    def get_validators(self, name: str) -> Optional[Dict[str, Validator]]:
         self.used_validators.add(name)
         specific_validators = self.validators.get(name)
         wildcard_validators = self.validators.get('*')
         if specific_validators or wildcard_validators:
             validators = (specific_validators or []) + (wildcard_validators or [])
             return {v.func.__name__: v for v in validators}
+        return None
 
-    def check_for_unused(self):
+    def check_for_unused(self) -> None:
         unused_validators = set(
             chain(
                 *[
@@ -90,8 +96,8 @@ class ValidatorGroup:
             )
 
 
-def extract_validators(namespace):
-    validators = {}
+def extract_validators(namespace: Dict[str, Any]) -> Dict[str, List[Validator]]:
+    validators: Dict[str, List[Validator]] = {}
     for var_name, value in namespace.items():
         validator_config = getattr(value, '__validator_config', None)
         if validator_config:
@@ -104,7 +110,7 @@ def extract_validators(namespace):
     return validators
 
 
-def inherit_validators(base_validators, validators):
+def inherit_validators(base_validators: ValidatorListDict, validators: ValidatorListDict) -> ValidatorListDict:
     for field, field_validators in base_validators.items():
         if field not in validators:
             validators[field] = []
@@ -112,14 +118,14 @@ def inherit_validators(base_validators, validators):
     return validators
 
 
-def get_validator_signature(validator):
+def get_validator_signature(validator: Any) -> ValidatorSignature:
     signature = inspect.signature(validator)
 
     # bind here will raise a TypeError so:
     # 1. we can deal with it before validation begins
     # 2. (more importantly) it doesn't get confused with a TypeError when executing the validator
     try:
-        if 'cls' in signature._parameters:
+        if 'cls' in signature._parameters:  # type: ignore
             if len(signature.parameters) == 2:
                 signature.bind(object(), 1)
                 return ValidatorSignature.CLS_JUST_VALUE