+++ pydantic/utils.py
@@ -6,7 +6,8 @@ from enum import Enum
 from functools import lru_cache
 from importlib import import_module
 from textwrap import dedent
-from typing import List, Pattern, Tuple, Type, _eval_type
+from typing import _eval_type  # type: ignore
+from typing import TYPE_CHECKING, Any, Dict, Generator, List, Optional, Pattern, Tuple, Type, Union
 
 from . import errors
 
@@ -16,25 +17,35 @@ except ImportError:
     email_validator = None
 
 try:
-    from typing import _TypingBase as typing_base
+    from typing import _TypingBase as typing_base  # type: ignore
 except ImportError:
-    from typing import _Final as typing_base
+    from typing import _Final as typing_base  # type: ignore
 
 try:
-    from typing import ForwardRef
+    from typing import ForwardRef  # type: ignore
 except ImportError:
     # python 3.6
     ForwardRef = None
 
+if TYPE_CHECKING:  # pragma: no cover
+    from .main import BaseModel  # noqa: F401
+
 if sys.version_info < (3, 7):
     from typing import Callable
+
+    AnyCallable = Callable[..., Any]
 else:
     from collections.abc import Callable
+    from typing import Callable as TypingCallable
+
+    AnyCallable = TypingCallable[..., Any]
+
 
 PRETTY_REGEX = re.compile(r'([\w ]*?) *<(.*)> *')
+AnyType = Type[Any]
 
 
-def validate_email(value) -> Tuple[str, str]:
+def validate_email(value: str) -> Tuple[str, str]:
     """
     Brutally simple email address validation. Note unlike most email address validation
     * raw ip address (literal) domain parts are not allowed.
@@ -49,10 +60,9 @@ def validate_email(value) -> Tuple[str, str]:
         raise ImportError('email-validator is not installed, run `pip install pydantic[email]`')
 
     m = PRETTY_REGEX.fullmatch(value)
+    name: Optional[str] = None
     if m:
         name, value = m.groups()
-    else:
-        name = None
 
     email = value.strip()
 
@@ -64,7 +74,7 @@ def validate_email(value) -> Tuple[str, str]:
     return name or email[: email.index('@')], email.lower()
 
 
-def _rfc_1738_quote(text):
+def _rfc_1738_quote(text: str) -> str:
     return re.sub(r'[:@/]', lambda m: '%{:X}'.format(ord(m.group(0))), text)
 
 
@@ -76,8 +86,8 @@ def make_dsn(
     host: str = None,
     port: str = None,
     name: str = None,
-    query: str = None,
-):
+    query: Dict[str, Any] = None,
+) -> str:
     """
     Create a DSN from from connection settings.
 
@@ -106,7 +116,7 @@ def make_dsn(
     return s
 
 
-def import_string(dotted_path):
+def import_string(dotted_path: str) -> Any:
     """
     Stolen approximately from django. Import a dotted module path and return the attribute/class designated by the
     last name in the path. Raise ImportError if the import fails.
@@ -123,7 +133,7 @@ def import_string(dotted_path):
         raise ImportError(f'Module "{module_path}" does not define a "{class_name}" attribute') from e
 
 
-def truncate(v, *, max_len=80):
+def truncate(v: str, *, max_len: int = 80) -> str:
     """
     Truncate a value and add a unicode ellipsis (three dots) to the end if it was too long
     """
@@ -136,7 +146,7 @@ def truncate(v, *, max_len=80):
     return v
 
 
-def display_as_type(v):
+def display_as_type(v: AnyType) -> str:
     if not isinstance(v, typing_base) and not isinstance(v, type):
         v = type(v)
 
@@ -155,19 +165,22 @@ def display_as_type(v):
         return str(v)
 
 
+ExcType = Type[Exception]
+
+
 @contextmanager
-def change_exception(raise_exc, *except_types):
+def change_exception(raise_exc: ExcType, *except_types: ExcType) -> Generator[None, None, None]:
     try:
         yield
     except except_types as e:
         raise raise_exc from e
 
 
-def clean_docstring(d):
+def clean_docstring(d: str) -> str:
     return dedent(d).strip(' \r\n\t')
 
 
-def list_like(v):
+def list_like(v: AnyType) -> bool:
     return isinstance(v, (list, tuple, set)) or inspect.isgenerator(v)
 
 
@@ -184,7 +197,7 @@ def validate_field_name(bases: List[Type['BaseModel']], field_name: str) -> None
 
 
 @lru_cache(maxsize=None)
-def url_regex_generator(*, relative: bool, require_tld: bool) -> Pattern:
+def url_regex_generator(*, relative: bool, require_tld: bool) -> Pattern[str]:
     """
     Url regex generator taken from Marshmallow library,
     for details please follow library source code:
@@ -214,30 +227,30 @@ def url_regex_generator(*, relative: bool, require_tld: bool) -> Pattern:
     )
 
 
-def lenient_issubclass(cls, class_or_tuple):
+def lenient_issubclass(cls: Any, class_or_tuple: Union[AnyType, Tuple[AnyType, ...]]) -> bool:
     return isinstance(cls, type) and issubclass(cls, class_or_tuple)
 
 
-def in_ipython():
+def in_ipython() -> bool:
     """
     Check whether we're in an ipython environment, including jupyter notebooks.
     """
     try:
-        __IPYTHON__
+        __IPYTHON__  # type: ignore
     except NameError:
         return False
     else:  # pragma: no cover
         return True
 
 
-def resolve_annotations(raw_annotations, module):
+def resolve_annotations(raw_annotations: Dict[str, AnyType], module_name: Optional[str]) -> Dict[str, AnyType]:
     """
     Partially taken from typing.get_type_hints.
 
     Resolve string or ForwardRef annotations into type objects if possible.
     """
-    if module:
-        base_globals = sys.modules[module].__dict__
+    if module_name:
+        base_globals: Optional[Dict[str, Any]] = sys.modules[module_name].__dict__
     else:
         base_globals = None
     annotations = {}
@@ -253,5 +266,5 @@ def resolve_annotations(raw_annotations, module):
     return annotations
 
 
-def is_callable_type(type_):
+def is_callable_type(type_: AnyType) -> bool:
     return type_ is Callable or getattr(type_, '__origin__', None) is Callable